:bug: Fix invalid data loading code in example

+++ example/client/agent/main.go
@@ -171,7 +171,7 @@ func load(path string) (ids []string, train, test [][]float32, err error) {
 		row, dim := int(dims[0]), int(dims[1])
 
 		// Gets the stored vector. All are represented as one-dimensional arrays.
-		vec := make([]float64, sp.SimpleExtentNPoints())
+		vec := make([]float32, sp.SimpleExtentNPoints())
 		if err := d.Read(&vec); err != nil {
 			return nil, err
 		}
